refactor(video): delete dead root twin, compute composition frames; single-source agent prompt (#127,#130)
