+++ src/App.tsx
@@ -12,20 +12,7 @@ import { DesignCard } from '@/components/catalog/DesignCard';
 import { AppliedDesignBanner } from '@/components/AppliedDesignBanner';
 import { ScrollToTop } from '@/components/ScrollToTop';
 import designs from '@/data/designs';
-
-const AGENT_PROMPT_TEMPLATE = (designUrl: string) => `Fetch the design system at: ${designUrl}
-
-Read the JSON, then follow the steps in agentInstructions.steps to apply this design system to my project:
-
-1. Set CSS custom properties from tokens.colors on :root (light) and .dark (dark mode)
-2. Set --radius from tokens.radius.default
-3. Load fonts by adding the Google Fonts URL from fonts.urls as a <link> tag
-4. Set font-family from tokens.typography.fontFamily
-5. Apply component styles from the components field (Tailwind class names for shadcn projects)
-6. Ensure focus states match accessibility.focusRing specification
-7. Test both light and dark modes
-
-Target framework: Tailwind CSS + shadcn/ui. For other frameworks, map token names to CSS custom properties semantically.`;
+import { buildAgentPrompt } from '@/lib/agentPrompt';
 
 const AGENT_STEPS = [
   {
@@ -64,7 +51,7 @@ const AGENTS = ['Claude Code', 'Cursor', 'Codex CLI', 'Windsurf', 'Copilot', 'Ge
 
 function getRandomPrompt() {
   const pick = designs[Math.floor(Math.random() * designs.length)];
-  return AGENT_PROMPT_TEMPLATE(pick.jsonUrl);
+  return buildAgentPrompt(pick.jsonUrl);
 }
 
 const PROMPT_EXAMPLE = getRandomPrompt();
