+++ video/src/index.tsx
@@ -1,12 +1,12 @@
 import { registerRoot } from "remotion";
 import { Composition } from "remotion";
-import { PromotionalVideo } from "./PromotionalVideo";
+import { PromotionalVideo, TOTAL_FRAMES } from "./PromotionalVideo";
 
 registerRoot(() => (
   <Composition
     id="promotional-video"
     component={PromotionalVideo}
-    durationInFrames={1020}
+    durationInFrames={TOTAL_FRAMES}
     fps={30}
     width={1920}
     height={1080}
